[erp] работа с кассой shtrih

+++ erp-logics/src/main/java/lsfusion/erp/region/by/machinery/cashregister/fiscalshtrih/UpdateDataTaxRate.java
@@ -1,12 +1,13 @@
 package lsfusion.erp.region.by.machinery.cashregister.fiscalshtrih;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 public class UpdateDataTaxRate implements Serializable {
     public Integer taxRateNumber;
-    public Double taxRateValue;
+    public BigDecimal taxRateValue;
 
-    public UpdateDataTaxRate(Integer taxRateNumber, Double taxRateValue) {
+    public UpdateDataTaxRate(Integer taxRateNumber, BigDecimal taxRateValue) {
         this.taxRateNumber = taxRateNumber;
         this.taxRateValue = taxRateValue;
     }
